refactor: move server to internal
