+++ app.go
@@ -2,8 +2,8 @@ package application
 
 import (
 	"echo-demo-project/internal/config"
-	"echo-demo-project/server"
-	"echo-demo-project/server/routes"
+	"echo-demo-project/internal/server"
+	"echo-demo-project/internal/server/routes"
 	"log"
 )
