+++ tests/auth_test.go
@@ -6,8 +6,8 @@ import (
 	"echo-demo-project/internal/models"
 	"echo-demo-project/internal/requests"
 	"echo-demo-project/internal/responses"
-	"echo-demo-project/server"
-	"echo-demo-project/server/handlers"
+	"echo-demo-project/internal/server"
+	"echo-demo-project/internal/server/handlers"
 	"echo-demo-project/internal/services/token"
 	"echo-demo-project/tests/helpers"
 	"encoding/json"
