+++ tests/helpers/server.go
@@ -2,7 +2,7 @@ package helpers
 
 import (
 	"echo-demo-project/config"
-	"echo-demo-project/server"
+	"echo-demo-project/internal/server"
 	"github.com/labstack/echo/v4"
 	"gorm.io/gorm"
 )
