+++ tests/register_test.go
@@ -3,8 +3,8 @@ package tests
 import (
 	"database/sql/driver"
 	"echo-demo-project/internal/requests"
-	"echo-demo-project/server"
-	"echo-demo-project/server/handlers"
+	"echo-demo-project/internal/server"
+	"echo-demo-project/internal/server/handlers"
 	"echo-demo-project/tests/helpers"
 	"net/http"
 	"testing"
